Rename constant to constants
